back_end/requirements.txt
flask
flask_cors
flask_sqlalchemy
psycopg2
PyJWT
bcrypt
flask-wtf
flask-security
email-validator
flask-migrate

